Move SocketId into its own module to prevent inline construction.
